Changes for DJUNITS 3.00.00. See OTS-319 in JIRA for explanation.
Right now, absolute and separate variables use the same unit. E.g., a Time and a Duration both use a TimeUnit. In classes that have to enforce that all components are of unit TimeUnit, Duration as well as Time can be used. This is not desirable. Therefore we need to split the TimeUnit, AngleUnit, LengthUnit and TemperatureUnit into an absolute and a relative version.

Absolute variables are variables with a relative offset to a defined value. E.g., 1 January of the year 1 can be used as the 'origin' for dates. Or 1 January 1970 for Unix times. For absolute time, this is called an Epoch (see https://en.wikipedia.org/wiki/Epoch_(reference_date)).
This means that several instances of the TimeUnit need to be made, as there are many epochs. This can be nicely done with the OffsetLinearScale that gives a ratio to use when transforming one scale into another as well as an offset. When subtracting a Unix system time from a date that is relative to 0001-01-01T00:00 AD, we know through the scale that the start of the Unix system time is 719162 days after the year-1 epoch. Using the scale, addition and subtraction operations can be carried out without making further changes.. 

An interesting unit is the TemperatureUnit. Here, we really messed up by mixing absolute temperature (including the offset) and relative temperature (which doesn't have an offset). We can say that another temperature is 10 degrees Celcius warmer than another temperature. Because of the offset, this was translated by the scale of the unit into 283.15 Kelvin... Ouch!

So, the TemperatureUnit has a LinearScale whereas the AbsoluteTemperatureUnit has an OffsetLinearScale.

the toAbs() and toRel() methods are removed. I don't understand how to translate 1-4-2017 at 12:00 to a relative value. The only thing that makes sense is to subtract a value (some kind offset) from the absolute value to get a relative value.

Removed the methods to multiply and divide absolute values with absolute values. How can I multiply a position on this planet with another position on the planet? Or a time with another time? Just doesn't make sense, so removed the methods and the corresponding tests.